Fix issue when banned user crashes server by adding more checks in routes. Adding error message for banned users.

+++ backend/routes/companyUserRoutes.js
@@ -85,7 +85,8 @@ companyRouter.get("/getCompanyJobs", async (req, res) => {
     try {
       const client = await pool.connect();
       //fetch user and check for banned status
-      await checkBanStatus(client, user_id, res);
+      const isBanned = await checkBanStatus(client, user_id, res);
+      if (isBanned) return;
       const result = await client.query(
         "SELECT * FROM jobs WHERE user_id = $1 ORDER BY job_id ASC",
         [user_id]
@@ -132,7 +133,8 @@ companyRouter.put("/editjob", async (req, res) => {
     try {
       const client = await pool.connect();
       //fetch user and check for banned status
-      await checkBanStatus(client, user_id, res);
+      const isBanned = await checkBanStatus(client, user_id, res);
+      if (isBanned) return;
       //fetch company name
       const company = await client.query(
         "SELECT fullname FROM users WHERE user_id = $1",
@@ -184,7 +186,8 @@ companyRouter.put("/canceljob", async (req, res) => {
     try {
       const client = await pool.connect();
       //fetch user and check for banned status
-      await checkBanStatus(client, user_id, res);
+      const isBanned = await checkBanStatus(client, user_id, res);
+      if (isBanned) return;
       await client.query("UPDATE jobs SET status = false WHERE job_id = $1", [
         job_id,
       ]);
@@ -212,7 +215,8 @@ companyRouter.delete("/deletejob", async (req, res) => {
     try {
       const client = await pool.connect();
       //fetch user and check for banned status
-      await checkBanStatus(client, user_id, res);
+      const isBanned = await checkBanStatus(client, user_id, res);
+      if (isBanned) return;
       //delete job
       await client.query("DELETE FROM jobs WHERE job_id = $1", [job_id]);
       const result = await client.query(
@@ -239,7 +243,8 @@ companyRouter.get("/getJobApplications", async (req, res) => {
     try {
       const client = await pool.connect();
       //fetch user and check for banned status
-      await checkBanStatus(client, user_id, res);
+      const isBanned = await checkBanStatus(client, user_id, res);
+      if (isBanned) return;
       const result = await client.query(
         "SELECT * FROM applications WHERE job_id = $1 ORDER BY user_id",
         [job_id]
@@ -278,7 +283,8 @@ companyRouter.put("/updateJobApplication", async (req, res) => {
     try {
       const client = await pool.connect();
       //fetch user and check for banned status
-      await checkBanStatus(client, userInfo.sub, res);
+      const isBanned = await checkBanStatus(client, userInfo.sub, res);
+      if (isBanned) return;
       await client.query(
         "UPDATE applications SET app_status = $1 WHERE job_id = $2 AND user_id = $3",
         [status, job_id, user_id]
@@ -323,7 +329,8 @@ companyRouter.post("/createMessage", async (req, res) => {
     try {
       const client = await pool.connect();
       //fetch user and check for banned status
-      await checkBanStatus(client, userInfo.sub, res);
+      const isBanned = await checkBanStatus(client, userInfo.sub, res);
+      if (isBanned) return;
       //select jobid from applications
       const job = await client.query(
         "SELECT * FROM jobs WHERE job_id = (SELECT job_id FROM applications WHERE app_id = $1)",
@@ -372,7 +379,8 @@ companyRouter.get("/getReplies", async (req, res) => {
     try {
       const client = await pool.connect();
       //fetch user and check for banned status
-      await checkBanStatus(client, user_id, res);
+      const isBanned = await checkBanStatus(client, user_id, res);
+      if (isBanned) return;
       //get all jobs_ids from company user
       const resultJobs = await client.query(
         "SELECT * FROM jobs WHERE user_id = $1",
@@ -446,7 +454,8 @@ companyRouter.delete("/deleteReply", async (req, res) => {
     try {
       const client = await pool.connect();
       //fetch user and check for banned status
-      await checkBanStatus(client, user_id, res);
+      const isBanned = await checkBanStatus(client, user_id, res);
+      if (isBanned) return;
       //check if the id is part of users jobs
       //get all jobs_ids from company user
       const resultJobs = await client.query(
